[Release 2.2.1] Added capability to define <query-defn> tags within <table> tags in SchemaDocuments. This capability allows the same type of "define and import" capability that was introduced for <dialog> and <statement> tags last week. Basically, here's what's allowed:
in schema.xml:

<table ...>
    <query-defn id="xyz">
          ... define...

in statements.xml

<xaf>
    <sql-statements ...>
    <table-query-defns table="*" prefix-table="yes"/>

By using the new <table-query-defns> tag in statements.xml, now each table's query defn will be "imported" into statements and automatically kept up to date. Unlike a <dialog> or <statement> tag which must be explicity defined inside <table> tags to be imported, even if a query definition is not explicity defined by a user a "default" query-defn that represents all the columns and a simple join will be automatically created.