Added the ability to auto-discover resources for the NavigationTree that are used by the skin.  The discovery process works as follow: 1. The skin is aware of the physical location of the files and the relative path      is needed to use the resources.  It invokes the NavigationTree (this class      was added to represent the ownerTree that used to be of type      NavigationPage) with the method discoverResources with those two      params. 2.  The NavigationTree then creates a Map of all the files found in the      specified directories maintining a hierarchical sturcture that resembles the      structure.xml 3.  The Skin then invokes resolveResources on NavigationTree which applies      inheritance to the resources so you only have to specify them to a certain      level if they are going to be the same from that level down.  In this process      if there is no entry in the resources for a particular id, it'll look at it's inmediate      parent and so on until it finds one that matches and it copies that as its own. 4.  The skin then is ready to get the image it needs by using the resources map.
From a files perspective, the default resources for the skin are place under
{app-root}/sparx/resources/skin/default/navigation
you can override or add images to the skin by placing them into
{app-root}/resources/skin/default/navigation
files in this directory will be automatically picked up by the discovery process.

Also added the ability to have a sub-heading for the pages that need it.  This
is a function of each page (NavigationPage) so a method was added to that
class (getSubHeading) that returns null by default.  The pages that desire to
display this, simply need to override this method and return a meaningful string
back to the skin.  If the skin gets a non-null >0 length string back when it calls
getSubHeading it will display the sub-heading accordingly.