Fixed bug that was causing the redirection not to happen when there was only one selection and it was setup to be hidden.
The problem: The SelectField.getHiddenControlHtml() only renders html
for selected chioces of the field.  Since there is no population of this field (it's
basically empty when created), the choice was never rendered.

The Solution: In the makeStateChanges() of this class, where is checks
whether it should be hidden if there is only one, it now also sets the value of
field to be the first one on the list (the only one in fact).  This solves the
problem.