* Added a new data type called lookup-with-freeform.  The best way to explain   what it does is through an example:
The single line in a schema of the form shown below...

<column name="language" type="lookup-with-freeform" size="64" enumeration-table="Language_Type" default="0" descr="The language for this person"/>

... will expand to the following multiple fields (with embedded UI field definitions)

<column name="language_id" lookupref="Language_Type" required="yes" descr="The language for this person">
  <field.select caption="Language Type"/>
</column>
<column name="language" type="text" size="64" descr="Descriptive name of an alternate language if choosing 'Other' for the field above">
  <field.text caption="Other Language Type">
    <conditional action="display" partner="language_id" js-expr="control.options[control.selectedIndex].text == 'Other'"/>
  </field.text>
</column>

The purpose is toe create one field that is a lookupref to an enumeration table
and a corresponding field that can be used when 'Other' is selected in the pulldown
for the lookup field.  In the example above, it creates a language_id field which is
filled with a numeric value from the Language_Type enumeration table.  It also
creates a language field which can be filled in manually.  In the UI, the language
field only appears if an option captioned 'Other' is present and selected in the
pulldown for language_id.

* Updated version number