Added the ability to define components for a page in the structure.xml.  This is an example of how to do it:
<page id="/a/b" caption="Summary">
                <components>
                    <layout columns="2"/>
                    <component cmd="statement,statement.name,-,-,detail" col-span="2"/>
                    <component cmd="statement,statement.name,-,-,record-viewer"/>
                    <component cmd="statement,statement.name,-,-,record-viewer" col-break="yes"/>
                    <component cmd="statement,statement.name,-,-,record-viewer"/>
                </components>
</page>

The scheme allows to place more col-breaks than specified in the layout and it
will adjust the layouts total number of columns.  A row-span is also allowed, with
the expected result.  If the cmd attribute is empty or the command cannot be resolved
that place in the table will be empty.

Right now, if a page has components defined, they will be shown unless a
component command is passed in the request.  Also, currently that page
cannot show the contents of a JSP.  Very soon that will be configurable both
through the xml and a request parameter.

This functionality is the last of the chain to be able to take care of most common
displays without the use of JSPs.  Using the component commands together
with this attribute, most pages can be defined within the structure.xml.

A document depicting all the functionality provided by navigation scheme will
be available soon.