[Version 2.2.24] Fixed a (perceived) problem with closing of connections. In many of the DAL routines in the Table object, the following code was being used:
if (rs != null)
{
    rs.getStatement().close();
    rs.close();
}

It turns out that according to the JDBC spec the rule is that if the statement gets closed, its associated resultset is also closed automatically. Most drivers don't complain if a resultset is closed, but some do. So, the new code is now specified as:

                if (rs != null)
                    rs.getStatement().close();

This will only close the statement which will automatically close the resultset assuming the driver follows the spec. Let me know if anyone sees connection leaks.