Implement Double Entry, Scan Entry, and Auto-Blur
On fixed-length fields like WBN, date, SSN, Auto-blur is a feature requested by DMIS users to automatically tab out of the field once the user has finished typing the characters spanning the field length.  This behavior is desgined to mimic the current DMIS behavior.

As an example, on a Date field, auto-blur will tab out of the field once the user finishes typing '03182003' or '03/18/2003'.  For a WBN field, auto-blur will automatically tab out once the user finishes typing 'LX12345', etc.

The developers is responsible for putting the correct tags in their corresponding dialogs where these features are required.

Here are examples/instructions:

------------------------------
<field.date query-field="drive_date" popup-calendar="no" double-entry="yes" required="yes" initial-focus="yes">
	<auto-blur length="8" exclude-expr="[/-]"/>
</field.date>

This will require a double entry on a date with auto-blur on the 8th character, excluding the slash and dash character.
------------------------------



------------------------------
<field.wbn name="wbn" required="yes" size="10" double-entry="yes">
	<scan-entry start-code="wbn" stop-code="wbn" partner="wbn_scanned"/>
	<auto-blur length="7"/>
</field.wbn>

This will require double entry on a WBN field, auto-blur on the 7th character. Also, this WBN field is scannable, with a start-code of 'wbn' and stop-code of 'wbn', just for kicks. If the field is scanned, and if you also specify a partner field named 'wbn_scanned' on the same dialog, the wbn_scanned field will be automatically populated with a 'yes' which you can act upon in your execute method.
------------------------------